requirements.txt
beautifulsoup4==4.11.1
nextcord==2.1.0
python-dotenv==0.20.0
requests==2.26.0
